Rename: src/serivce/team으로 묶음
team.validation
team.parser
